+++ src/Header/ContentType.php
@@ -4,31 +4,28 @@ declare(strict_types = 1);
 namespace Innmind\Http\Header;
 
 use Innmind\Http\Header as HeaderInterface;
-use Innmind\Immutable\Sequence;
+use Innmind\MediaType\MediaType;
+use Innmind\Immutable\{
+    Sequence,
+    Str,
+};
 
 /**
  * @psalm-immutable
  */
 final class ContentType implements HeaderInterface
 {
-    public function __construct(
-        private ContentTypeValue $content,
+    private function __construct(
+        private MediaType $content,
     ) {
     }
 
     /**
      * @psalm-pure
      */
-    public static function of(
-        string $type,
-        string $subType,
-        Parameter ...$parameters,
-    ): self {
-        return new self(new ContentTypeValue(
-            $type,
-            $subType,
-            ...$parameters,
-        ));
+    public static function of(MediaType $content): self
+    {
+        return new self($content);
     }
 
     #[\Override]
@@ -43,7 +40,7 @@ final class ContentType implements HeaderInterface
         return $this->header()->values();
     }
 
-    public function content(): ContentTypeValue
+    public function content(): MediaType
     {
         return $this->content;
     }
@@ -56,6 +53,33 @@ final class ContentType implements HeaderInterface
 
     private function header(): Header
     {
-        return new Header('Content-Type', $this->content);
+        $mediaType = new MediaType(
+            $this->content->topLevel(),
+            $this->content->subType(),
+            $this->content->suffix(),
+        );
+        $parameters = Str::of(';')->join(
+            $this
+                ->content
+                ->parameters()
+                ->map(static fn($parameter) => new Parameter\Parameter( // to make sure it's of the HTTP format
+                    $parameter->name(),
+                    $parameter->value(),
+                ))
+                ->map(static fn($parameter) => $parameter->toString()),
+        );
+
+        $content = $mediaType->toString();
+
+        if (!$parameters->empty()) {
+            $content .= ';';
+        }
+
+        $content .= $parameters->toString();
+
+        return new Header(
+            'Content-Type',
+            new Value\Value($content),
+        );
     }
 }