+++ tests/Header/ContentTypeTest.php
@@ -6,45 +6,51 @@ namespace Tests\Innmind\Http\Header;
 use Innmind\Http\{
     Header\ContentType,
     Header,
-    Header\ContentTypeValue,
-    Header\Parameter
 };
+use Innmind\MediaType\MediaType;
 use Innmind\Immutable\Sequence;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\Attributes\DataProvider;
 
 class ContentTypeTest extends TestCase
 {
     public function testInterface()
     {
-        $h = new ContentType(
-            $ct = new ContentTypeValue(
-                'text',
-                'html',
-                new Parameter\Parameter('charset', 'UTF-8'),
-            ),
+        $h = ContentType::of(
+            $ct = MediaType::of('text/html; charset="UTF-8"'),
         );
 
         $this->assertInstanceOf(Header::class, $h);
         $this->assertSame('Content-Type', $h->name());
         $v = $h->values();
         $this->assertInstanceOf(Sequence::class, $v);
-        $this->assertSame($ct, $v->find(static fn() => true)->match(
-            static fn($first) => $first,
-            static fn() => null,
-        ));
+        $this->assertSame($ct, $h->content());
         $this->assertSame('Content-Type: text/html;charset=UTF-8', $h->toString());
-        $this->assertSame('text/html;charset=UTF-8', $h->content()->toString());
+        $this->assertSame('text/html; charset=UTF-8', $h->content()->toString());
     }
 
-    public function testOf()
+    #[DataProvider('invalids')]
+    public function testReturnNothingWhenInvalidContentTypeValue($type, $sub)
     {
-        $header = ContentType::of(
-            'text',
-            'html',
-            new Parameter\Parameter('charset', 'UTF-8'),
+        $this->assertNull(
+            MediaType::maybe("$type/$sub")
+                ->map(ContentType::of(...))
+                ->match(
+                    static fn($header) => $header,
+                    static fn() => null,
+                ),
         );
+    }
 
-        $this->assertInstanceOf(ContentType::class, $header);
-        $this->assertSame('Content-Type: text/html;charset=UTF-8', $header->toString());
+    public static function invalids(): array
+    {
+        return [
+            ['*', '*'],
+            ['*', 'octet-stream'],
+            ['text', '*'],
+            ['foo/bar', ''],
+            ['foo', 'bar+suffix'],
+            ['foo', 'bar, level=1'],
+        ];
     }
 }